+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### 🚀 New Cards
 
 ### 🐛 Defect Fixes
+- **Issue #93:** - Colors and Typography are not saving correctly in the yaml for slider, analog clock, and gauge cards.
+
 
 
 ## v26.3.1